Refactoring: Renaming prefereces to preferences filename
