+++ lib/mercadopago-support.js
@@ -4,7 +4,7 @@ var requestManager = require('./request-manager');
 var paymentModule = require('./resources/payment');
 var collectionsModule = require('./resources/collections');
 var preapprovalModule = require('./resources/preapproval');
-var preferencesModule = require('./resources/prefereces');
+var preferencesModule = require('./resources/preferences');
 var ipnModule = require('./resources/ipn');
 
 module.exports = function () {
