+++ lib/mercadopago.js
@@ -2,7 +2,7 @@ var configurations = require('./configurations');
 var backwardSupport = require('./mercadopago-support')();
 var utilsModule = require('./utils');
 var paymentModule = require('./resources/payment');
-var preferencesModule = require('./resources/prefereces');
+var preferencesModule = require('./resources/preferences');
 var preapprovalModule = require('./resources/preapproval');
 var merchantOrdersModule = require('./resources/merchantOrders');
 var moneyRequestsModule = require('./resources/moneyRequests');
