+++ lib/resources/preferences.js
